feat: add task management functions including create, update, delete, duplicate, and tag tasks
