+++ src/actions/tasks/update-task.ts
@@ -1,0 +1,72 @@
+"use server"
+
+import { auth } from "@/auth"
+import { db } from "@/lib/db"
+import { updateTaskSchema } from "@/validations/task"
+import {
+  parseDatetime,
+  parseTags,
+  revalidateTaskPaths,
+  type ActionResult,
+} from "./shared"
+
+export async function updateTask(
+  taskId: string,
+  formData: FormData
+): Promise<ActionResult<void>> {
+  const session = await auth()
+  if (!session?.user?.id) {
+    return { success: false, error: "Unauthorized" }
+  }
+
+  const tags = parseTags(formData)
+  const dueDateRaw = formData.get("dueDate") as string | null
+  const raw = {
+    title: formData.get("title") || undefined,
+    description: formData.get("description") || undefined,
+    priority: formData.get("priority") || undefined,
+    status: formData.get("status") || undefined,
+    tags: tags.length > 0 ? tags : undefined,
+    projectId: formData.get("projectId") || undefined,
+    dueDate: dueDateRaw !== null ? (dueDateRaw.trim() || null) : undefined,
+  }
+
+  const parsed = updateTaskSchema.safeParse(raw)
+  if (!parsed.success) {
+    return { success: false, error: parsed.error.errors[0].message }
+  }
+
+  const existing = await db.task.findUnique({
+    where: { id: taskId, userId: session.user.id },
+    select: { id: true, projectId: true },
+  })
+
+  if (!existing) {
+    return { success: false, error: "Task not found" }
+  }
+
+  if (parsed.data.projectId && parsed.data.projectId !== existing.projectId) {
+    const project = await db.project.findUnique({
+      where: { id: parsed.data.projectId, userId: session.user.id },
+      select: { id: true },
+    })
+    if (!project) {
+      return { success: false, error: "Project not found" }
+    }
+  }
+
+  await db.task.update({
+    where: { id: taskId, userId: session.user.id },
+    data: {
+      ...parsed.data,
+      dueDate: parsed.data.dueDate === null
+        ? null
+        : parsed.data.dueDate
+          ? parseDatetime(parsed.data.dueDate)
+          : undefined,
+    },
+  })
+
+  revalidateTaskPaths()
+  return { success: true, data: undefined }
+}
